Add files via upload
Tkinter is a powerful library in Python that is used for creating graphical user interfaces (GUIs). One of the most important features of Tkinter is the ability to create multiple window frames, which can be used to organize and display different parts of your application. In this blog post, we will explore how to create multiple window frames with Tkinter and how to use them effectively in your Python applications.